data_processing/requirements.txt
jupyter
elevenlabs
deepgram-sdk
python-dotenv